HydrologyPlanetesimals

Documentation for HydrologyPlanetesimals.

HydrologyPlanetesimals.StaticParameters

HydrologyPlanetesimals.Q_radiogenic

HydrologyPlanetesimals.apply_insulating_boundary_conditions!

HydrologyPlanetesimals.assemble_hydromechanical_lse!

HydrologyPlanetesimals.calculate_radioactive_heating

HydrologyPlanetesimals.compute_Aϕ!

HydrologyPlanetesimals.compute_basic_node_properties!

HydrologyPlanetesimals.compute_displacement_timestep

HydrologyPlanetesimals.compute_fluid_velocities!

HydrologyPlanetesimals.compute_marker_properties!

HydrologyPlanetesimals.compute_nodal_adjustment!

HydrologyPlanetesimals.compute_p_node_properties!

HydrologyPlanetesimals.compute_stress_strainrate!

HydrologyPlanetesimals.compute_vx_node_properties!

HydrologyPlanetesimals.compute_vy_node_properties!

HydrologyPlanetesimals.define_markers!

HydrologyPlanetesimals.distance

HydrologyPlanetesimals.finalize_plastic_iteration_pass!

HydrologyPlanetesimals.fix_weights

HydrologyPlanetesimals.get_viscosities_stresses_density_gradients!

HydrologyPlanetesimals.interpolate_to_grid!

HydrologyPlanetesimals.interpolate_to_marker!

HydrologyPlanetesimals.kphi

HydrologyPlanetesimals.ktotal

HydrologyPlanetesimals.positive_max!

HydrologyPlanetesimals.process_hydromechanical_solution!

HydrologyPlanetesimals.recompute_bulk_viscosity!

HydrologyPlanetesimals.run_simulation

HydrologyPlanetesimals.setup_dynamic_simulation_parameters

HydrologyPlanetesimals.setup_staggered_grid_geometry

HydrologyPlanetesimals.setup_staggered_grid_properties

HydrologyPlanetesimals.simulation_loop

HydrologyPlanetesimals.symmetrize_p_node_observables!

HydrologyPlanetesimals.total

HydrologyPlanetesimals.StaticParameters — Type

Static parameters: Grids, markers, switches, constants, etc. which remain constant throughout the simulation.

hr_al::Bool

radioactive heating from 26Al active Default: true

hr_fe::Bool

radioactive heating from 60Fe active Default: true

xsize::Float64

horizontal model size [m] Default: 140000.0

ysize::Float64

vertical model size [m] Default: 140000.0

xcenter::Float64

horizontal center of model Default: xsize / 2

ycenter::Float64

vertical center of model Default: ysize / 2

Nx::Int64

basic grid resolution in x direction (horizontal) Default: 141

Ny::Int64

basic grid resolution in y direction (vertical) Default: 141

Nx1::Int64

Vx, Vy, P grid resolution in x direction (horizontal) Default: Nx + 1

Ny1::Int64

Vx/Vy/P grid resolution in y direction (vertical) Default: Ny + 1

dx::Float64

horizontal grid step [m] Default: xsize / (Nx - 1)

dy::Float64

vertical grid step [m] Default: ysize / (Ny - 1)

jmin_basic::Int64

minimum assignable basic grid index in x direction Default: 1

imin_basic::Int64

minimum assignable basic grid index in y direction Default: 1

jmax_basic::Int64

maximum assignable basic grid index in x direction Default: Nx - 1

imax_basic::Int64

maximum assignable basic grid index in y direction Default: Ny - 1

jmin_vx::Int64

minimum assignable Vx grid index in x direction Default: 1

imin_vx::Int64

minimum assignable Vx grid index in y direction Default: 1

jmax_vx::Int64

maximum assignable Vx grid index in x direction Default: Nx - 1

imax_vx::Int64

maximum assignable Vx grid index in y direction Default: Ny

jmin_vy::Int64

minimum assignable Vy grid index in x direction Default: 1

imin_vy::Int64

minimum assignable Vy grid index in y direction Default: 1

jmax_vy::Int64

maximum assignable Vy grid index in x direction Default: Nx

imax_vy::Int64

maximum assignable Vy grid index in y direction Default: Ny - 1

jmin_p::Int64

minimum assignable P grid index in x direction Default: 1

imin_p::Int64

minimum assignable P grid index in y direction Default: 1

jmax_p::Int64

maximum assignable P grid index in x direction Default: Nx

imax_p::Int64

maximum assignable P grid index in y direction Default: Ny

rplanet::Float64

planetary radius [m] Default: 50000.0

rcrust::Float64

crust radius [m] Default: 48000.0

psurface::Float64

surface pressure [Pa] Default: 1000.0

Nxmc::Int64

number of markers per cell in horizontal direction Default: 4

Nymc::Int64

number of markers per cell in vertical direction Default: 4

Nxm::Int64

marker grid resolution in horizontal direction Default: (Nx - 1) * Nxmc

Nym::Int64

marker grid resolution in vertical direction Default: (Ny - 1) * Nymc

dxm::Float64

marker grid step in horizontal direction Default: xsize / Nxm

dym::Float64

marker grid step in vertical direction Default: ysize / Nym

start_marknum::Int64

number of markers at start Default: Nxm * Nym

G::Float64

gravitational constant [m^3kg^-1s^-2] Default: 6.672e-11

pscale::Float64

scaled pressure Default: 1.0e23 / dx

rhosolidm::StaticArrays.SVector{3, Float64}

solid Density [kg/m^3] Default: [3300.0, 3300.0, 1.0]

rhofluidm::StaticArrays.SVector{3, Float64}

fluid density [kg/m^3] Default: [7000.0, 7000.0, 1000.0]

etasolidm::StaticArrays.SVector{3, Float64}

solid viscosity [Pa*s] Default: [1.0e16, 1.0e16, 1.0e14]

etasolidmm::StaticArrays.SVector{3, Float64}

molten solid viscosity [Pa*s] Default: [1.0e14, 1.0e14, 1.0e14]

etafluidm::StaticArrays.SVector{3, Float64}

fluid viscosity [Pa*s] Default: [0.01, 0.01, 1.0e12]

etafluidmm::StaticArrays.SVector{3, Float64}

molten fluid viscosity [Pa*s] Default: [0.01, 0.01, 1.0e12]

rhocpsolidm::StaticArrays.SVector{3, Float64}

solid volumetric heat capacity [kg/m^3] Default: [3.3e6, 3.3e6, 3.0e6]

rhocpfluidm::StaticArrays.SVector{3, Float64}

fluid volumetric heat capacity [kg/m^3] Default: [7.0e6, 7.0e6, 3.0e6]

alphasolidm::StaticArrays.SVector{3, Float64}

solid thermal expansion [1/K] Default: [3.0e-5, 3.0e-5, 0.0]

alphafluidm::StaticArrays.SVector{3, Float64}

fluid thermal expansion [1/K] Default: [5.0e-5, 5.0e-5, 0.0]

ksolidm::StaticArrays.SVector{3, Float64}

solid thermal conductivity [W/m/K] Default: [3.0, 3.0, 3000.0]

kfluidm::StaticArrays.SVector{3, Float64}

fluid thermal conductivity [W/m/K] Default: [50.0, 50.0, 3000.0]

start_hrsolidm::StaticArrays.SVector{3, Float64}

solid radiogenic heat production [W/m^3] Default: [0.0, 0.0, 0.0]

start_hrfluidm::StaticArrays.SVector{3, Float64}

fluid radiogenic heat production [W/m^3] Default: [0.0, 0.0, 0.0]

gggsolidm::StaticArrays.SVector{3, Float64}

solid shear modulus [Pa] Default: [1.0e10, 1.0e10, 1.0e10]

frictsolidm::StaticArrays.SVector{3, Float64}

solid friction coefficient Default: [0.6, 0.6, 0.0]

cohessolidm::StaticArrays.SVector{3, Float64}

solid compressive strength [Pa] Default: [1.0e8, 1.0e8, 1.0e8]

tenssolidm::StaticArrays.SVector{3, Float64}

solid tensile strength [Pa] Default: [6.0e7, 6.0e7, 6.0e7]

kphim0::StaticArrays.SVector{3, Float64}

standard permeability [m^2] Default: [1.0e-13, 1.0e-13, 1.0e-17]

tkm0::StaticArrays.SVector{3, Float64}

initial temperature [K] Default: [300.0, 300.0, 273.0]

etaphikoef::Float64

Coefficient to compute compaction viscosity from shear viscosity Default: 0.0001

t_half_al::Float64

26Al half life [s] Default: 717000 * 31540000

tau_al::Float64

26Al decay constant Default: thalfal / log(2)

ratio_al::Float64

initial ratio of 26Al and 27Al isotopes Default: 5.0e-5

E_al::Float64

E 26Al [J] Default: 5.047e-13

f_al::Float64

26Al atoms/kg Default: 1.9e23

t_half_fe::Float64

60Fe half life [s] Default: 2620000 * 31540000

tau_fe::Float64

60Fe decay constant Default: thalffe / log(2)

ratio_fe::Float64

initial ratio of 60Fe and 56Fe isotopes Default: 1.0e-6

E_fe::Float64

E 60Fe [J] Default: 4.34e-13

f_fe::Float64

60Fe atoms/kg Default: 1.957e24

tmsilicate::Float64

silicate melting temperature [K] Default: 1.0e6

tmiron::Float64

iron melting temperature [K] Default: 1273

phim0::Float64

standard Fe fraction [porosity] Default: 0.2

phimin::Float64

min porosity Default: 0.0001

phimax::Float64

max porosity Default: 1 - phimin

bcleft::Float64

mechanical boundary condition left Default: -1

bcright::Float64

mechanical boundary condition right Default: -1

bctop::Float64

mechanical boundary condition top Default: -1

bcbottom::Float64

mechanical boundary condition bottom Default: -1

bcfleft::Float64

hydraulic boundary condition left Default: -1

bcfright::Float64

hydraulic boundary condition right Default: -1

bcftop::Float64

hydraulic boundary condition top Default: -1

bcfbottom::Float64

hydraulic boundary condition bottom Default: -1

strainrate::Float64

shortening strain rate Default: 0.0

vxleft::Float64

x extension/shortening velocity left Default: (strainrate * xsize) / 2

vxright::Float64

x extension/shortening velocity right Default: (-strainrate * xsize) / 2

vytop::Float64

y extension/shortening velocity top Default: (-strainrate * ysize) / 2

vybottom::Float64

y extension/shortening velocity bottom Default: (strainrate * ysize) / 2

nname::String

mat filename Default: madcph_

savematstep::Int64

.mat storage periodicity Default: 50

dtelastic::Float64

Maximal computational timestep [s] Default: 1.0e11

dtkoef::Float64

Coefficient to decrease computational timestep Default: 2

dtkoefup::Float64

Coefficient to increase computational timestep Default: 1.1

dtstep::Int64

Number of iterations before changing computational timestep Default: 200

dxymax::Float64

Max marker movement per time step [grid steps] Default: 0.05

vpratio::Float64

Weight of averaged velocity for moving markers Default: 1 / 3

DTmax::Float64

Max temperature change per time step [K] Default: 20

dsubgridt::Float64

Subgrid temperature diffusion parameter Default: 0

dsubgrids::Float64

Subgrid stress diffusion parameter Default: 0

yearlength::Float64

length of year [s] Default: 365.25 * 24 * 3600

start_time::Float64

Time sum (start) [s] Default: 1.0e6yearlength

endtime::Float64

Time sum (end) [s] Default: 15 * 1000000 * yearlength

etamin::Float64

Lower viscosity cut-off [Pa s] Default: 1.0e12

etamax::Float64

Upper viscosity cut-off [Pa s] Default: 1.0e23

nplast::Int64

Number of plastic iterations Default: 100000

visstep::Int64

Periodicity of visualization Default: 1

yerrmax::Float64

Tolerance level for yielding error() Default: 100.0

etawt::Float64

Weight for old viscosity Default: 0

dphimax::Float64

max porosity ratio change per time step Default: 0.01

start_step::Int64

starting timestep Default: 1

nsteps::Int64

number of timesteps to run Default: 30000

HydrologyPlanetesimals.Q_radiogenic — Method

Compute radiogenic heat production of isotope mixture.

Q_radiogenic(f, ratio, E, tau, time)

Details

- f: fraction of radioactive matter [atoms/kg]
- ratio: initial ratio of radioactive to non-radioactive isotopes
- E: heat energy [J]
- tau: exp decay mean lifetime ``\tau=\frac{t_{1/2}}{\log{2}}`` [s]
- time: time elapsed since start of radioactive decay [s]

Returns

- Q: radiogenic heat production [W/kg]

HydrologyPlanetesimals.apply_insulating_boundary_conditions! — Method

Apply insulating boundary conditions to given array.

[x x x x x x [a a b c d d

x a b c d x a a b c d d

x e f g h x -> e e f g h h

x x x x x x] e e f g h h]

Details

- t: array to apply insulating boundary conditions to

Returns

- nothing

HydrologyPlanetesimals.assemble_hydromechanical_lse! — Method

Assemble hydromechanical system of equations.

assemble_hydromechanical_lse!(ETAcomp, ETAPcomp, SXYcomp, SXXcomp, SYYcomp, dRHOXdx, dRHOXdy, dRHOYdx, dRHOYdy, RHOX, RHOY, RHOFX, RHOFY, RX, RY, ETAPHI, BETTAPHI, PHI, gx, gy, pr0, pf0, dt, L, R, sp)

Details

- ETAcomp: computational viscosity at basic nodes
- ETAPcomp: computational viscosity at P nodes
- SXYcomp: computational previous XY stress at basic nodes
- SXXcomp: computational previous XX stress at P nodes
- SYYcomp: computational previous YY stress at P nodes
- dRHOXdx: total density gradient in x direction at Vx nodes
- dRHOXdy: total density gradient in y direction at Vx nodes
- dRHOYdx: total density gradient in x direction at Vy nodes
- dRHOYdy: total density gradient in y direction at Vy nodes
- RHOX: total density at Vx nodes
- RHOY: total density at Vy nodes
- RHOFX: fluid density at Vx nodes
- RHOFY: fluid density at Vy nodes
- RX: ηfluid/Kϕ at Vx nodes
- RY: ηfluid/Kϕ at Vy nodes
- ETAPHI: bulk viscosity at P nodes
- BETTAPHI: bulk compressibility at P nodes
- PHI: porosity at P nodes
- gx: x gravitational acceleration at Vx nodes
- gy: y gravitational acceleration at Vy nodes
- pr0: previous total pressure at P nodes
- pf0: previous fluid pressure at P nodes
- dt: time step
- L: ExtendableSparse matrix to store LHS coefficients
- R: vector to store RHS coefficients
- sp: simulation parameters

Returns

- nothing

HydrologyPlanetesimals.calculate_radioactive_heating — Method

Compute radiogenic heat production of 26Al and 60Fe isotopes.

calculate_radioactive_heating(timesum, sp)

Details

- timesum: time elapsed since initial conditions at start of simulation
- sp: static simulation parameters

Returns

- hrsolidm: radiogenic heat production of 26Al [W/m^3]
- hrfluidm: radiogenic heat production of 60Fe [W/m^3]

HydrologyPlanetesimals.compute_Aϕ! — Method

Compute porosity coefficient Aϕ = Dln[(1-ϕ)/ϕ]/Dt

compute_Aϕ!(APHI, ETAPHI, BETTAPHI, PHI, pr, pf, pr0, pf0, dt)

Details

In

- ETAPHI: bulk viscosity at P Nodes
- BETTAPHI: bulk compressibility at P nodes
- PHI: porosity at P Nodes
- pr: total pressure at P nodes
- pf: fluid pressure at P nodes
- pr0: previous step total pressure at P nodes
- pf0: previous step fluid pressure at P nodes
- dt: time step

Out

- APHI: porosity coefficient at P nodes

Returns

- aphimax: maximum absolute porosity coefficient

HydrologyPlanetesimals.compute_basic_node_properties! — Method

Compute properties of basic nodes based on interpolation arrays.

compute_basic_node_properties!(ETA0SUM, ETASUM, GGGSUM, SXYSUM, COHSUM, TENSUM, FRISUM, WTSUM, ETA0, ETA, GGG, SXY0, COH, TEN, FRI, YNY)

Details

- ETA0SUM: ETA0 interpolation array
- ETASUM: ETA interpolation array
- GGGSUM: GGG interpolation array
- SXYSUM: SXY interpolation array
- COHSUM: COH interpolation array
- TENSUM: TEN interpolation array
- FRISUM: FRI interpolation array
- WTSUM: WT interpolation array
- ETA0: ETA0 basic node array
- ETA: ETA basic node array
- GGG: GGG basic node array
- SXY0: SXY basic node array
- COH: COH basic node array
- TEN: TEN basic node array
- FRI: FRI basic node array
- YNY: YNY basic node array

Returns

- nothing

HydrologyPlanetesimals.compute_displacement_timestep — Method

Compute displacement time step.

compute_displacement_timestep(vx, vy, vxf, vyf, dt, dx, dy, dxymax, aphimax, dphimax)

Details

- vx: solid vx velocity at Vx nodes
- vy: solid vy velocity at Vy nodes
- vxf: fluid vx velocity at Vx nodes
- vyf: fluid vy velocity at Vy nodes
- dt: current time step
- dx: x-grid horizontal spacing
- dy: y-grid vertical spacing
- dxymax: maximum allowed grid spacing movement of markers per time step
- aphimax: maximum observed porosity coefficient
- dphimax: maximum allowed porosity ratio change per time step

Returns

- dtm: displacement time step

HydrologyPlanetesimals.compute_fluid_velocities! — Method

Compute current fluid velocity.

compute_fluid_velocities!(PHIX, PHIY, qxD, qyD, vx, vy, vxf, vyf, sp)

Details

In

- PHIX: porosity at Vx nodes
- PHIY: porosity at Vy nodes
- qxD: qx-Darcy flux at Vx nodes
- qyD: qy-Darcy flux at Vy nodes
- vx: solid velocity at Vx nodes
- vy: solid velocity at Vy nodes
- sp: simulation parameters

Out

- vxf: fluid vx velocity at Vx nodes
- vyf: fluid vy velocity at Vy nodes

Returns

- nothing

HydrologyPlanetesimals.compute_marker_properties! — Method

Compute properties of given marker and save them to corresponding arrays.

compute_marker_properties!(m, tm, tkm, rhototalm, rhocptotalm, etasolidcur, etafluidcur, etatotalm, hrtotalm, ktotalm, tkm_rhocptotalm, etafluidcur_inv_kphim, hrsolidm, hrfluidm, phim, sp)

Details

- m: marker number
- tm: array of type of markers
- tkm: array of temperature of markers
- rhototalm: array of total density of markers
- rhocptotalm: array of total volumetric heat capacity of markers
- etasolidcur: array of solid viscosity of markers
- etafluidcur: array of fluid viscosity of markers
- etatotalm: array of total viscosity of markers
- hrtotalm: array of total radiogenic heat production of markers
- ktotalm: array of total thermal conductivity of markers
- tkm_rhocptotalm: array of total thermal energy of markers
- etafluidcur_inv_kphim: array of (fluid viscosity)/permeability of markers
- phim: array of porosity of markers
- hrsolidm: vector of radiogenic heat production of solid materials
- hrfluidm: vector of radiogenic heat production of fluid materials
- sp: static simulation parameters

Returns

- nothing

HydrologyPlanetesimals.compute_nodal_adjustment! — Method

Compute nodal adjustment and return plastic iterations completeness status.

compute_nodal_adjustment!(ETA, ETA0, ETA5, GGG, SXX, SXY, pr, pf, COH, TEN, FRI, SIIB, siiel, prB, pfB, syieldc, syieldt, syield, etapl, YNY, YNY5, YERRNOD, DSY, YNPL, DDD, dt, iplast, sp)

Details

- ETA: viscoplastic viscosity at basic nodes
- ETA0: previous time step viscoplastic viscosity at basic nodes
- ETA5: plastic iterations viscoplastic viscosity at basic nodes
- GGG: shear modulus at basic nodes
- SXX: σ′xx at P nodes
- SXY: σxy at basic nodes
- pr: total pressure at P nodes
- pf: fluid pressure at P nodes
- COH: compressive strength at basic nodes 
- TEN: tensile strength at basic nodes 
- FRI: friction at basic nodes
- SIIB: second stress invariant at basic nodes
- siiel: second invariant for purely elastic stress buildup at basic nodes 
- prB: interpolated total pressure at basic nodes 
- pfB: interpolated fluid pressure at basic nodes 
- syieldc: confined fractures yielding stress at basic nodes 
- syieldt: tensile fractures yielding stress at basic nodes 
- syield: non-negative maximum yielding stress at basic nodes
- etapl: stress-based viscoplastic viscosity at basic nodes 
- YNY: plastic yielding status at basic nodes 
- YNY5: plastic iterations plastic yielding status at basic nodes
- YERRNOD: vector of summed yielding errors of nodes over plastic iterations
- DSY: (SIIB-syield) at basic nodes
- YNPL: plastic iterations plastic yielding status at basic nodes
- DDD: plastic iterations (SIIB-syield)² at basic nodes
- dt: time set
- iplast: plastic iteration step 
- sp: static simulation parameters

Returns

- plastic_iterations_complete: true if plastic iterations complete

HydrologyPlanetesimals.compute_p_node_properties! — Method

Compute properties of P nodes based on interpolation arrays.

compute_p_node_properties!(RHOSUM, RHOCPSUM, ALPHASUM, ALPHAFSUM, HRSUM, GGGPSUM, SXXSUM, TKSUM, PHISUM, WTPSUM, RHO, RHOCP, ALPHA, ALPHAF, HR, GGGP, SXX0, tk1, PHI, BETTAPHI)

Details

- GGGPSUM: GGGP interpolation array
- SXX0SUM: SXX0 interpolation array
- RHOSUM: RHO interpolation array
- RHOCPSUM: RHOCP interpolation array
- ALPHASUM: ALPHA interpolation array
- ALPHAFSUM: ALPHAF interpolation array
- HRSUM: HR interpolation array
- PHISUM: PHI interpolation array
- TKSUM: TK interpolation array
- WTPSUM: WTP interpolation array
- GGGP: GGGP P node array
- SXX0: SXX0 P node array
- RHO: RHO P node array
- RHOCP: RHOCP P node array
- ALPHA: ALPHA P node array
- ALPHAF: ALPHAF P node array
- HR: HR P node array
- PHI: PHI P node array
- BETTAPHI: BETTAPHI P node array
- tk1: tk1 P node array

Returns

- nothing

HydrologyPlanetesimals.compute_stress_strainrate! — Method

Compute stress, stress change, and strain rate components.

compute_stress_strainrate!(vx, vy, ETA, GGG, ETAP, GGGP, SXX0, SXY0, EXX, EXY, SXX, SXY, DSXX, DSXY, EII, SII, dtm, sp)

Details

In

- vx: solid vx velocity at Vx nodes
- vy: solid vy velocity at Vy nodes
- ETA: viscosity at basic nodes
- GGG: shear modulus at basic nodes
- ETAP: viscosity at P nodes
- GGGP: shear modulus at P nodes
- SXX0: previous time step σ₀′xx at P nodes
- SXY0: previous time step σ₀xy at basic nodes
- dtm: displacement time step
- sp: simulation parameters

Out

- EXX: ϵxx at P nodes
- EXY: ϵxy at basic nodes
- SXX: σ′xx P nodes
- SXY: σxy at basic nodes
- DSXX: stress change Δσ′xx at P nodes
- DSXY: stress change Δσxy at basic nodes
- EII: second strain rate invariant ϵᴵᴵ at P nodes
- SII: second stress invariant σᴵᴵ at P nodes

Returns

    - nothing

HydrologyPlanetesimals.compute_vx_node_properties! — Method

Compute properties of Vx nodes based on interpolation arrays.

compute_vx_node_properties!(RHOXSUM, RHOFXSUM, KXSUM, PHIXSUM, RXSUM, WTXSUM, RHOX, RHOFX, KX, PHIX, RX)

Details

- RHOXSUM: RHOX interpolation array
- RHOFXSUM: RHOFX interpolation array
- KXSUM: KX interpolation array
- PHIXSUM: PHIX interpolation array
- RXSUM: RX interpolation array
- WTXSUM: WTX interpolation array
- RHOX: RHOX Vx node array
- RHOFX: RHOFX Vx node array
- KX: KX Vx node array
- PHIX: PHIX Vx node array
- RX: RX Vx node array

Returns

- nothing

HydrologyPlanetesimals.compute_vy_node_properties! — Method

Compute properties of Vy nodes based on interpolation arrays.

compute_vy_node_properties!(RHOYSUM, RHOFYSUM, KYSUM, PHIYSUM, RYSUM, WTYSUM, RHOY, RHOFY, KY, PHIY, RY)

Details

- RHOYSUM: RHOY interpolation array
- RHOFYSUM: RHOFY interpolation array
- KYSUM: KY interpolation array
- PHIYSUM: PHIY interpolation array
- RYSUM: RY interpolation array
- WTYSUM: WTY interpolation array
- RHOY: RHOY Vy node array
- RHOFY: RHOFY Vy node array
- KY: KY Vy node array
- PHIY: PHIY Vy node array
- RY: RY Vy node array

Returns

- nothing

HydrologyPlanetesimals.define_markers! — Method

Define initial set of markers according to model parameters

define_markers!(xm, ym, tm, phim, etavpm, rhototalm, rhocptotalm, etatotalm, hrtotalm, ktotalm, etafluidcur, tkm, inv_gggtotalm, fricttotalm, cohestotalm, tenstotalm, rhofluidcur, alphasolidcur, alphafluidcur, sp)

Details

- xm: array of x coordinates of markers
- ym: array of y coordinates of markers
- tm: array of material type of markers
- phim: array of porosity of markers
- etavpm: array of matrix viscosity of markers
- rhototalm: array of total density of markers
- rhocptotalm: array of total volumetric heat capacity of markers
- etatotalm: array of total viscosity of markers
- hrtotalm: array of total radiogenic heat production of markers
- ktotalm: array of total thermal conductivity of markers
- etafluidcur: array of fluid viscosity of markers
- tkm: array of temperature of markers 
- inv_gggtotalm: array of inverse of total shear modulus of markers
- fricttotalm: array of total friction coefficient of markers
- cohestotalm: array of total compressive strength of markers
- tenstotalm: array of total tensile strength of markers
- rhofluidcur: array of fluid density of markers
- alphasolidcur: array of solid thermal expansion coefficient of markers
- alphafluidcur: array of fluid thermal expansion coefficient of markers
- sp: static simulation parameters

Returns

- nothing

HydrologyPlanetesimals.distance — Method

Calculate Euclidean distance between two point coordinates.

distance(x1, y1, x2, y2)

Details

- x1: x-coordinate of point 1 [m]
- y1: y-coordinate of point 1 [m]
- x2: x-coordinate of point 2 [m]
- y2: y-coordinate of point 2 [m]

Returns

- Euclidean distance between point 1 and point 2 [m]

HydrologyPlanetesimals.finalize_plastic_iteration_pass! — Method

Decide next pass plastic iteration time step, viscoplastic viscosity, and basic node yielding status.

finalize_plastic_iteration_pass!(ETA, ETA5, ETA00, YNY, YNY5, YNY00, dt, dtkoef, dtstep, iplast)

Details:

- ETA: viscoplastic viscosity at basic nodes
- ETA5: plastic iterations viscoplastic viscosity at basic nodes
- ETA00: previous time step viscoplastic viscosity at basic nodes
- YNY: plastic yielding status at basic nodes 
- YNY5: plastic iterations plastic yielding status at basic nodes
- YNY00: previous time step plastic yielding status at basic nodes
- dt: current time step
- dtkoef: coefficient by which to decrease computational time step
- dtstep: minimum of plastic iterations before adjusting time step
- iplast: current plastic iteration counter

Returns

- dt: adjusted next time step

HydrologyPlanetesimals.fix_weights — Method

Compute top and left grid nodes indices and bilinear interpolation weigths to nearest four grid nodes for given (x, y) position and grid axes.

fix_weights(x, y, x_axis, y_axis, dx, dy, jmin, jmax, imin, imax)

Details

- x: x-position [m]
- y: y-position [m]
- x_axis: x-grid reference axis array [m]
- y_axis: y-grid reference axis array [m]
- dx: x-grid axis mesh width [m]
- dy: y-grid axis mesh width [m]
- jmin: minimum assignable index on x-grid axis (basic/Vx/Vy/P)
- jmax: maximum assignable index on x-grid axis (basic/Vx/Vy/P)
- imin: minimum assignable index on y-grid axis (basic/Vx/Vy/P)
- imax: maximum assignable index on y-grid axis (basic/Vx/Vy/P)

Returns

- i: top (with reference to y) node index on y-grid axis
- j: left (with reference to x) node index on x-grid axis
- bilinear_weights: vector of 4 bilinear interpolation weights to
  nearest four grid nodes:
    [wtmij  : i  , j   node,
    wtmi1j : i+1, j   node,
    wtmij1 : i  , j+1 node,
    wtmi1j1: i+1, j+1 node]

HydrologyPlanetesimals.get_viscosities_stresses_density_gradients! — Method

Compute viscosities, stresses, and density gradients for hydromechanical solver.

get_viscosities_stresses_density_gradients!(ETA, ETAP, GGG, GGGP, SXY0, SXX0, RHOX, RHOY, dx, dy, dt, Nx, Ny, Nx1, Ny1, ETAcomp, ETAPcomp, SXYcomp, SXXcomp, SYYcomp, dRHOXdx, dRHOXdy, dRHOYdx, dRHOYdy)

Details

In

- ETA: viscoplastic viscosity at basic nodes
- ETAP: viscosity at P nodes
- GGG: shear modulus at basic nodes
- GGGP: shear modulus at P nodes
- SXY0: σ₀xy XY stress at basic nodes
- SXX0: σ₀′xx XX stress at P nodes
- RHOX: density at Vx nodes
- RHOY: density at Vy nodes
- dx: horizontal grid spacing
- dy: vertical grid spacing
- dt: time step
- Nx: number of horizontal basic grid points
- Ny: number of vertical basic grid points
- Nx1: number of horizontal Vx/Vy/P grid points
- Ny1: number of vertical Vx/Vy/P grid points

Out

- ETAcomp: computational viscosity at basic nodes
- ETAPcomp: computational viscosity at P nodes
- SXYcomp: previous XY stresses at basic nodes
- SXXcomp: previous XX stresses at P nodes
- SYYcomp: previous YY stresses at P nodes
- dRHOXdx: density gradient at Vx nodes in x direction
- dRHOXdy: density gradient at Vx nodes in y direction
- dRHOYdx: density gradient at Vy nodes in x direction
- dRHOYdy: density gradient at Vy nodes in y direction

Returns

- nothing

HydrologyPlanetesimals.interpolate_to_grid! — Method

Interpolate a property to neareast four nodes on a given grid location using given bilinear interpolation weights.

Details

- i: top (with reference to y) node index on vertical y-grid axis
- j: left (with reference to x) node index on horizontal x-grid axis
- weights: vector of 4 bilinear interpolation weights to
  nearest four grid nodes:
    [wtmij  : i  , j   node,
    wtmi1j : i+1, j   node,
    wtmij1 : i  , j+1 node,
    wtmi1j1: i+1, j+1 node]
- property: property to be interpolated to grid using weights
- grid: threaded grid array on which to interpolate property

Returns

- nothing

HydrologyPlanetesimals.interpolate_to_marker! — Method

Interpolate a property from nearest the four nodes on a given grid to a marker.

Details

- m: number of marker to interpolate to
- i: top (with reference to y) node index on vertical y-grid axis
- j: left (with reference to x) node index on horizontal x-grid axis
- weights: vector of 4 bilinear interpolation weights to
nearest four grid nodes:
    [wtmij  : i  , j   node,
    wtmi1j : i+1, j   node,
    wtmij1 : i  , j+1 node,
    wtmi1j1: i+1, j+1 node]
- marker_property: marker property array into which to interpolate
- property_grid: grid whose property is to be interpolated to marker
- m: marker

HydrologyPlanetesimals.kphi — Method

Compute iron porosity-dependent permeability.

kphi(kphim0, phim, phim0)

Details

- kphim0: standard permeability [m^2]
- phim: actual (marker) porosity
- phim0: standard iron fraction (porosity)

Returns

- kphim: iron porosity-dependent permeability [m^2]

HydrologyPlanetesimals.ktotal — Method

Compute total thermal conductivity of two-phase material.

ktotal(ksolid, kfluid, phi)

Details

- ksolid: solid thermal conductivity [W/m/K]
- kfluid: fluid thermal conductivity [W/m/K]
- phi: fraction of solid

Returns

- ktotal: total thermal conductivity of mixed phase [W/m/K]

HydrologyPlanetesimals.positive_max! — Method

Compare two arrays of identical sizes element-wise and fill a third array with the larger value if positive and zero otherwise.

positive_max!(A, B, C)

Details

- A: first array
- B: second array
- C: result array

Returns

- nothing

HydrologyPlanetesimals.process_hydromechanical_solution! — Method

Process hydromechanical solution vector to output physical observables.

process_hydromechanical_solution!(S, vx, vy, pr, qxD, qyD, pf, pscale, Nx1, Ny1)

Details

- S: hydromechanical solution vector
- vx: solid velocity at Vx nodes
- vy: solid velocity at Vy nodes
- pr: total pressure at P nodes
- qxD: qx-Darcy flux at Vx nodes
- qyD: qy-Darcy flux at Vy nodes
- pf: fluid pressure at P nodes
- pscale: scaled pressure
- Nx1: number of Vx/Vy/P nodes in horizontal x-direction
- Ny1: number of Vx/Vy/P nodes in vertical y-direction

Returns

- nothing

HydrologyPlanetesimals.recompute_bulk_viscosity! — Method

Recompute bulk viscosity at P nodes.

Details

- ETA: viscoplastic viscosity at basic nodes
- ETAP: viscosity at P Nodes
- ETAPHI: bulk viscosity at P Nodes
- PHI: porosity at P Nodes
- etaphikoef: coefficient: shear viscosity -> compaction viscosity

Returns

- nothing

HydrologyPlanetesimals.run_simulation — Function

Runs the simulation with the given parameters.

Details

- xsize: size of the domain in horizontal x direction [m]
- ysize: size of the domain in vertical y direction [m]
- rplanet: radius of the planet [m]
- rcrust: radius of the crust [m]
- Nx: number of basic grid nodes in horizontal x direction
- Ny: number of basic grid nodes in vertical y direction
- Nxmc: initial number of markers per cell in horizontal x direction
- Nymc: initial number of markers per cell in vertical y direction

Returns

- exit code

HydrologyPlanetesimals.setup_dynamic_simulation_parameters — Method

setup_dynamic_simulation_parameters(sp)

Details

- sp: static simulation parameters

Returns

- timestep: simulation starting time step count
- dt: simulation initial computational time step 
- timesum: simulation starting time
- marknum: initial number of markers
- hrsolidm: initial radiogenic heat production solid phase
- hrfluidm: initial radiogenic heat production fluid phase
- YERRNOD: vector of summed yielding errors of nodes over plastic iterations

HydrologyPlanetesimals.setup_staggered_grid_geometry — Method

Set up staggered grid geometry with basic, Vx, Vy, and P nodes.

setup_staggered_grid_geometry(sp)

Details

- sp: static simulation parameters

Returns

- x:
- y:
- xvx:
- yvx:
- xvy:
- yvy:
- xp:
- yp:

HydrologyPlanetesimals.setup_staggered_grid_properties — Method

Set up staggered grid properties for basic, Vx, Vy, and P nodes.

setup_staggered_grid_properties(sp; randomized)

Details

- sp: static simulation parameters
- randomized: fill in random values for grid properties instead of zeros

Returns

-

HydrologyPlanetesimals.simulation_loop — Method

Main simulation loop: run calculations with timestepping.

simulation_loop(sp)

Details

- markers: arrays containing all marker properties
- sp: static simulation parameters

Returns

- nothing

HydrologyPlanetesimals.symmetrize_p_node_observables! — Method

Apply symmetry to P node observables.

symmetrize_p_node_observables!(SXX, APHI, PHI, pr, pf, ps, Nx, Ny, Nx1, Ny1)

Details

- SXX: σ′xx at P nodes
- APHI: Aϕ = Dln[(1-ϕ)/ϕ]/Dt at P nodes
- PHI: porosity at P nodes
- pr: total pressure at P nodes
- pf: fluid pressure at P nodes
- ps: solid pressure at P nodes
- Nx: number of basic nodes in horizontal x direction
- Ny: number of basic nodes in vertical y direction
- Nx1: number of Vx/Vy/P nodes in horizontal x direction
- Ny1: number of Vx/Vy/P nodes in vertical y direction

Returns

- nothing

HydrologyPlanetesimals.total — Method

Compute convex combination of fluid and solid properties to get total property.

total(solid, fluid, phi)

Details

- fluid: fluid properties
- solid: solid properties
- phi: fraction of fluid

Returns

- total: computed total property