HydrologyPlanetesimals

Documentation for HydrologyPlanetesimals.

HydrologyPlanetesimals.Q_radiogenic

HydrologyPlanetesimals.add_vrk4

HydrologyPlanetesimals.apply_insulating_boundary_conditions!

HydrologyPlanetesimals.apply_subgrid_stress_diffusion!

HydrologyPlanetesimals.apply_subgrid_temperature_diffusion!

HydrologyPlanetesimals.assemble_hydromechanical_lse!

HydrologyPlanetesimals.backtrace_pressures_rk4!

HydrologyPlanetesimals.calculate_radioactive_heating

HydrologyPlanetesimals.compute_Aϕ!

HydrologyPlanetesimals.compute_adiabatic_heating!

HydrologyPlanetesimals.compute_basic_node_properties!

HydrologyPlanetesimals.compute_displacement_timestep

HydrologyPlanetesimals.compute_fluid_velocities!

HydrologyPlanetesimals.compute_gravity_solution!

HydrologyPlanetesimals.compute_marker_properties!

HydrologyPlanetesimals.compute_nodal_adjustment!

HydrologyPlanetesimals.compute_p_node_properties!

HydrologyPlanetesimals.compute_rotation_rate!

HydrologyPlanetesimals.compute_shear_heating!

HydrologyPlanetesimals.compute_stress_strainrate!

HydrologyPlanetesimals.compute_velocities!

HydrologyPlanetesimals.compute_vx_node_properties!

HydrologyPlanetesimals.compute_vy_node_properties!

HydrologyPlanetesimals.define_markers!

HydrologyPlanetesimals.distance

HydrologyPlanetesimals.etatotal_rocks

HydrologyPlanetesimals.finalize_plastic_iteration_pass!

HydrologyPlanetesimals.fix

HydrologyPlanetesimals.fix_distances

HydrologyPlanetesimals.fix_weights

HydrologyPlanetesimals.get_viscosities_stresses_density_gradients!

HydrologyPlanetesimals.grid_average

HydrologyPlanetesimals.grid_vector

HydrologyPlanetesimals.interpolate_add_to_grid!

HydrologyPlanetesimals.interpolate_add_to_marker!

HydrologyPlanetesimals.interpolate_to_marker!

HydrologyPlanetesimals.kphi

HydrologyPlanetesimals.ktotal

HydrologyPlanetesimals.marker_to_basic_nodes!

HydrologyPlanetesimals.marker_to_p_nodes!

HydrologyPlanetesimals.marker_to_vx_nodes!

HydrologyPlanetesimals.marker_to_vy_nodes!

HydrologyPlanetesimals.move_markers_rk4!

HydrologyPlanetesimals.parse_commandline

HydrologyPlanetesimals.perform_thermal_iterations!

HydrologyPlanetesimals.positive_max!

HydrologyPlanetesimals.process_hydromechanical_solution!

HydrologyPlanetesimals.recompute_bulk_viscosity!

HydrologyPlanetesimals.reduce_add_3darray!

HydrologyPlanetesimals.replenish_markers!

HydrologyPlanetesimals.reset_interpolated_properties

HydrologyPlanetesimals.run_simulation

HydrologyPlanetesimals.save_state

HydrologyPlanetesimals.setup_dynamic_simulation_parameters

HydrologyPlanetesimals.setup_gravitational_lse

HydrologyPlanetesimals.setup_hydromechanical_lse

HydrologyPlanetesimals.setup_interpolated_properties

HydrologyPlanetesimals.setup_marker_geometry_helpers

HydrologyPlanetesimals.setup_marker_properties

HydrologyPlanetesimals.setup_marker_properties_helpers

HydrologyPlanetesimals.setup_staggered_grid_properties

HydrologyPlanetesimals.setup_staggered_grid_properties_helpers

HydrologyPlanetesimals.setup_thermal_lse

HydrologyPlanetesimals.simulation_loop

HydrologyPlanetesimals.symmetrize_p_node_observables!

HydrologyPlanetesimals.total

HydrologyPlanetesimals.update_marker_population_geometry!

HydrologyPlanetesimals.update_marker_porosity!

HydrologyPlanetesimals.update_marker_stress!

HydrologyPlanetesimals.update_marker_temperature!

HydrologyPlanetesimals.update_marker_viscosity!

HydrologyPlanetesimals.ηᶠcur_inv_kᵠ

HydrologyPlanetesimals.Q_radiogenic — Method

Compute radiogenic heat production of isotope mixture.

Q_radiogenic(f, ratio, E, tau, time)

Details

- f: fraction of radioactive matter [atoms/kg]
- ratio: initial ratio of radioactive to non-radioactive isotopes
- E: heat energy [J]
- tau: exp decay mean lifetime ``\tau=\frac{t_{1/2}}{\log{2}}`` [s]
- time: time elapsed since start of radioactive decay [s]

Returns

- Q: radiogenic heat production [W/kg]

HydrologyPlanetesimals.add_vrk4 — Method

Add a RK4 stage velocity to the RK4 velocity vector.

add_vrk4(vrk4, v, rk)

Details:

- vrk4: current RK4 velocity vector
- v: RK4 velocity of stage `rk` to be added to velocity vector
- rk: RK4 stage number

Returns

- vrk4: updated RK4 velocity vector

HydrologyPlanetesimals.apply_insulating_boundary_conditions! — Method

Apply insulating boundary conditions to given array:

[x x x x x x; x a b c d x; x e f g h x; x x x x x x]

becomes

[a a b c d d; a a b c d d; e e f g h h; e e f g h h]

Details

- t: array to apply insulating boundary conditions to

Returns

- nothing

HydrologyPlanetesimals.apply_subgrid_stress_diffusion! — Method

Apply xy (basic grid) and xx (P grid) subgrid stress diffusion to markers.

apply_subgrid_stress_diffusion!(xm, ym, tm, inv_gggtotalm, sxxm, sxym, SXX0, SXY0, DSXX, DSXY, SXXSUM, SXYSUM, WTPSUM, WTSUM, dtm, marknum)

Details

- xm: marker x-coordinates
- ym: marker y-coordinates
- tm: marker type
- inv_gggtotalm: inverse of total shear modulus of markers
- sxxm: marker σ′xx [Pa]
- sxym: marker σxy [Pa]
- SXX0: σ₀′xx at P nodes [1/s]
- SXY0: σ₀xy at basic nodes [1/s]
- DSXX: stress change Δσ′xx at P nodes
- DSXY: stress change Δσxy at basic nodes
- SXXSUM: interpolation of SXX at P nodes
- SXYSUM: interpolation of SXY at basic nodes
- WTPSUM: interpolation weights at P nodes
- WTSUM: interpolation weights at basic nodes
- dtm: displacement time step
- marknum: total number of markers in use

Returns

- nothing

HydrologyPlanetesimals.apply_subgrid_temperature_diffusion! — Method

Apply subgrid temperature diffusion to markers.

apply_subgrid_temperature_diffusion!(xm, ym, tm, tkm, phim, tk1, DT, TKSUM, RHOCPSUM, dtm, marknum)

Details

- xm: marker x-coordinates 
- ym: marker y-coordinates
- tm: marker type
- tkm: marker temperature
- phim: marker porosity
- tk1: current temperature at P nodes
- DT: temperature change at P nodes
- TKSUM: interpolation of TK at P nodes 
- RHOCPSUM: interpolation of RHOCP at P nodes
- dtm: displacement time step
- marknum: total number of markers in use

Returns

- nothing

HydrologyPlanetesimals.assemble_hydromechanical_lse! — Method

Assemble hydromechanical system of equations.

assemble_hydromechanical_lse!(ETA, ETAP, GGG, GGGP, SXY0, SXX0, RHOX, RHOY, RHOFX, RHOFY, RX, RY, ETAPHI, BETTAPHI, PHI, gx, gy, pr0, pf0, dt, R)

Details

- ETA: viscosity at basic nodes
- ETAP: viscosity at P nodes
- GGG: shear modulus at basic nodes
- GGGP: shear modulus at P nodes
- SXY0: previous XY stress at basic nodes
- SXX0: previous XX stress at P nodes
- RHOX: total density at Vx nodes
- RHOY: total density at Vy nodes
- RHOFX: fluid density at Vx nodes
- RHOFY: fluid density at Vy nodes
- RX: ηfluid/Kϕ at Vx nodes
- RY: ηfluid/Kϕ at Vy nodes
- ETAPHI: bulk viscosity at P nodes
- BETTAPHI: bulk compressibility at P nodes
- PHI: porosity at P nodes
- gx: x gravitational acceleration at Vx nodes
- gy: y gravitational acceleration at Vy nodes
- pr0: previous total pressure at P nodes
- pf0: previous fluid pressure at P nodes
- dt: time step
- R: vector to store RHS coefficients

Returns

- L: LHS coefficient matrix

HydrologyPlanetesimals.backtrace_pressures_rk4! — Method

Backtrack pressure nodes using classic Runge-Kutta integration (RK4) to update total, solid, and fluid pressure under consideration of two-phase flow velocities.

backtrace_pressures_rk4!(pr, pr0, ps, ps0, pf, pf0, vx, vy, vxf, vyf, dtm)

Details

- pr:
- pr0:
- ps:
- ps0:
- pf:
- pf0:
- vx: 
- vy: 
- vxf: 
- vyf:

Returns

- nothing

HydrologyPlanetesimals.calculate_radioactive_heating — Method

Compute radiogenic heat production of 26Al and 60Fe isotopes.

calculate_radioactive_heating(al, fe, timesum)

Details

- al: true if radioactive isotope 26Al is present
- fe: true if radioactive isotope 60Fe is present
- timesum: time elapsed since initial conditions at start of simulation

Returns

- hrsolidm: radiogenic heat production of 26Al [W/m^3]
- hrfluidm: radiogenic heat production of 60Fe [W/m^3]

HydrologyPlanetesimals.compute_Aϕ! — Method

Compute porosity coefficient Aϕ = Dln[(1-ϕ)/ϕ]/Dt

compute_Aϕ!(APHI, ETAPHI, BETTAPHI, PHI, pr, pf, pr0, pf0, dt)

Details

In

- ETAPHI: bulk viscosity at P Nodes
- BETTAPHI: bulk compressibility at P nodes
- PHI: porosity at P Nodes
- pr: total pressure at P nodes
- pf: fluid pressure at P nodes
- pr0: previous step total pressure at P nodes
- pf0: previous step fluid pressure at P nodes
- dt: time step

Out

- APHI: porosity coefficient at P nodes

Returns

- aphimax: maximum absolute porosity coefficient

HydrologyPlanetesimals.compute_adiabatic_heating! — Method

Compute adiabatic heating based on basic (temperature) and P grids.

compute_adiabatic_heating!(HA, tk1, ALPHA, ALPHAF, PHI, vx, vy, vxf, vyf, ps, pf)

Details

- HA: adiabatic heating at P nodes
- tk1: previous temperature at P nodes
- ALPHA: thermal expansion coefficient at P nodes
- ALPHAF: fluid thermal expansion coefficient at P nodes
- PHI: porosity at P nodes
- vx: solid vx-velocity at Vx nodes
- vy: solid vy-velocity at Vy nodes
- vxf: fluid vx-velocity at Vx nodes
- vyf: fluid vy-velocity at Vy nodes
- ps: solid pressure at P nodes
- pf: fluid pressure at P nodes

Returns

- nothing

HydrologyPlanetesimals.compute_basic_node_properties! — Method

Compute properties of basic nodes based on interpolation arrays.

compute_basic_node_properties!(ETA0SUM, ETASUM, GGGSUM, SXYSUM, COHSUM, TENSUM, FRISUM, WTSUM, ETA0, ETA, GGG, SXY0, COH, TEN, FRI, YNY)

Details

- ETA0SUM: ETA0 interpolation array
- ETASUM: ETA interpolation array
- GGGSUM: GGG interpolation array
- SXYSUM: SXY interpolation array
- COHSUM: COH interpolation array
- TENSUM: TEN interpolation array
- FRISUM: FRI interpolation array
- WTSUM: WT interpolation array
- ETA0: ETA0 basic node array
- ETA: ETA basic node array
- GGG: GGG basic node array
- SXY0: SXY basic node array
- COH: COH basic node array
- TEN: TEN basic node array
- FRI: FRI basic node array
- YNY: YNY basic node array

Returns

- nothing

HydrologyPlanetesimals.compute_displacement_timestep — Method

Compute displacement time step.

compute_displacement_timestep(vx, vy, vxf, vyf, dt, aphimax)

Details

- vx: solid vx velocity at Vx nodes
- vy: solid vy velocity at Vy nodes
- vxf: fluid vx velocity at Vx nodes
- vyf: fluid vy velocity at Vy nodes
- dt: current time step
- dxymax: maximum allowed grid spacing movement of markers per time step
- aphimax: maximum observed porosity coefficient
- dphimax: maximum allowed porosity ratio change per time step

Returns

- dtm: displacement time step

HydrologyPlanetesimals.compute_fluid_velocities! — Method

Compute current fluid velocities.

compute_fluid_velocities!(PHIX, PHIY, qxD, qyD, vx, vy, vxf, vyf)

Details

In

- PHIX: porosity at Vx nodes
- PHIY: porosity at Vy nodes
- qxD: qx-Darcy flux at Vx nodes
- qyD: qy-Darcy flux at Vy nodes
- vx: solid velocity at Vx nodes
- vy: solid velocity at Vy nodes

Out

- vxf: fluid vx velocity at Vx nodes
- vyf: fluid vy velocity at Vy nodes

Returns

- nothing

HydrologyPlanetesimals.compute_gravity_solution! — Method

Compute gravity solution in P nodes to obtain gravitational accelerations gx for Vx nodes, gy for Vy nodes.

compute_gravity_solution!(SP, RP, RHO, FI, gx, gy)

Details

- SP: solution vector
- RP: right hand side vector
- RHO: density at P nodes
- FI: gravity potential at P nodes
- gx: x gravitational acceleration at Vx nodes
- gy: y gravitational acceleration at Vy nodes

Returns

nothing

HydrologyPlanetesimals.compute_marker_properties! — Method

Compute properties of given marker and save them to corresponding arrays.

compute_marker_properties!(m, tm, tkm, rhototalm, rhocptotalm, etatotalm, hrtotalm, ktotalm, tkm_rhocptotalm, etafluidcur_inv_kphim, hrsolidm, hrfluidm, phim)

Details

- m: marker number
- tm: type of markers
- tkm: temperature of markers
- rhototalm: total density of markers
- rhocptotalm: total volumetric heat capacity of markers
- etatotalm: total viscosity of markers
- hrtotalm: total radiogenic heat production of markers
- ktotalm: total thermal conductivity of markers
- tkm_rhocptotalm: total thermal energy of markers
- etafluidcur_inv_kphim: (fluid viscosity)/permeability of markers
- hrsolidm: vector of radiogenic heat production of solid materials
- hrfluidm: vector of radiogenic heat production of fluid materials
- phim: porosity of markers

Returns

- nothing

HydrologyPlanetesimals.compute_nodal_adjustment! — Method

Compute nodal adjustment and return plastic iterations completeness status.

compute_nodal_adjustment!(ETA, ETA0, ETA5, GGG, SXX, SXY, pr, pf, COH, TEN, FRI, YNY, YNY5, YERRNOD, DSY, dt, iplast)

Details

- ETA: viscoplastic viscosity at basic nodes
- ETA0: previous time step viscoplastic viscosity at basic nodes
- ETA5: plastic iterations viscoplastic viscosity at basic nodes
- GGG: shear modulus at basic nodes
- SXX: σ′xx at P nodes
- SXY: σxy at basic nodes
- pr: total pressure at P nodes
- pf: fluid pressure at P nodes
- COH: compressive strength at basic nodes 
- TEN: tensile strength at basic nodes 
- FRI: friction at basic nodes
- YNY: plastic yielding status at basic nodes 
- YNY5: plastic iterations plastic yielding status at basic nodes
- YERRNOD: vector of summed yielding errors of nodes over plastic iterations
- DSY: (SIIB-syield) at basic nodes
- dt: time set
- iplast: plastic iteration step

Returns

- plastic_iterations_complete: true if plastic iterations complete

HydrologyPlanetesimals.compute_p_node_properties! — Method

Compute properties of P nodes based on interpolation arrays.

compute_p_node_properties!(RHOSUM, RHOCPSUM, ALPHASUM, ALPHAFSUM, HRSUM, GGGPSUM, SXXSUM, TKSUM, PHISUM, WTPSUM, RHO, RHOCP, ALPHA, ALPHAF, HR, GGGP, SXX0, tk1, PHI, BETTAPHI)

Details

- GGGPSUM: GGGP interpolation array
- SXX0SUM: SXX0 interpolation array
- RHOSUM: RHO interpolation array
- RHOCPSUM: RHOCP interpolation array
- ALPHASUM: ALPHA interpolation array
- ALPHAFSUM: ALPHAF interpolation array
- HRSUM: HR interpolation array
- PHISUM: PHI interpolation array
- TKSUM: TK interpolation array
- WTPSUM: WTP interpolation array
- GGGP: GGGP P node array
- SXX0: SXX0 P node array
- RHO: RHO P node array
- RHOCP: RHOCP P node array
- ALPHA: ALPHA P node array
- ALPHAF: ALPHAF P node array
- HR: HR P node array
- PHI: PHI P node array
- BETTAPHI: BETTAPHI P node array
- tk1: tk1 P node array

Returns

- nothing

HydrologyPlanetesimals.compute_rotation_rate! — Method

Compute rotatation rate in basic nodes based on velocity derivatives at Vx and Vy nodes.

compute_rotation_rate!(vx, vy, wyx)

Details

- vx: solid vx-velocity at Vx nodes
- vy: solid vy-velocity at Vy nodes
- wyx: rotation rate at basic nodes

Returns

- nothing

HydrologyPlanetesimals.compute_shear_heating! — Method

Compute shear heating based on basic (temperature) and P grids.

compute_shear_heating!(HS, ETA, SXY, ETAP, SXX, RX, RY, qxD, qyD, PHI, ETAPHI, pr, pf)

Details

- HS: shear heating
- ETA: viscoplastic viscosity at basic nodes
- SXY: σ₀xy XY stress at basic nodes
- ETAP: viscosity at P nodes
- SXX: σ₀xy XY stress at basic nodes
- RX: ηfluid/Kϕ at Vx nodes
- RY: ηfluid/Kϕ at Vy nodes
- qxD: qx-Darcy flux at Vx nodes
- qyD: qy-Darcy flux at Vy nodes
- PHI: porosity at P nodes
- ETAPHI: bulk viscosity at P nodes
- pr: total pressure at P nodes
- pf: fluid pressure at P nodes

Returns

- nothing

HydrologyPlanetesimals.compute_stress_strainrate! — Method

Compute stress, stress change, and strain rate components.

compute_stress_strainrate!(vx, vy, ETA, GGG, ETAP, GGGP, SXX0, SXY0, EXX, EXY, SXX, SXY, DSXX, DSXY, EII, SII, dtm)

Details

In

- vx: solid vx velocity at Vx nodes
- vy: solid vy velocity at Vy nodes
- ETA: viscosity at basic nodes
- GGG: shear modulus at basic nodes
- ETAP: viscosity at P nodes
- GGGP: shear modulus at P nodes
- SXX0: previous time step σ₀′xx at P nodes
- SXY0: previous time step σ₀xy at basic nodes
- dtm: displacement time step

Out

- EXX: ϵxx at P nodes
- EXY: ϵxy at basic nodes
- SXX: σ′xx P nodes
- SXY: σxy at basic nodes
- DSXX: stress change Δσ′xx at P nodes
- DSXY: stress change Δσxy at basic nodes
- EII: second strain rate invariant ϵᴵᴵ at P nodes
- SII: second stress invariant σᴵᴵ at P nodes

Returns

    - nothing

HydrologyPlanetesimals.compute_velocities! — Method

Compute solid velocities, fluid velocities at P nodes.

compute_velocities!(vx, vy, vxf, vyf, vxp, vyp, vxpf, vypf)

Details

- vx: solid vx-velocity at Vx nodes
- vy: solid vy-velocity at Vy nodes
- vxf: fluid vx-velocity at Vx nodes
- vyf: fluid vy-velocity at Vy nodes
- vxp: solid vx-velocity at P nodes
- vyp: solid vy-velocity at P nodes
- vxpf: fluid vx-velocity at P nodes
- vypf: fluid vy-velocity at P nodes

Returns

- nothing

HydrologyPlanetesimals.compute_vx_node_properties! — Method

Compute properties of Vx nodes based on interpolation arrays.

compute_vx_node_properties!(RHOXSUM, RHOFXSUM, KXSUM, PHIXSUM, RXSUM, WTXSUM, RHOX, RHOFX, KX, PHIX, RX)

Details

- RHOXSUM: RHOX interpolation array
- RHOFXSUM: RHOFX interpolation array
- KXSUM: KX interpolation array
- PHIXSUM: PHIX interpolation array
- RXSUM: RX interpolation array
- WTXSUM: WTX interpolation array
- RHOX: RHOX Vx node array
- RHOFX: RHOFX Vx node array
- KX: KX Vx node array
- PHIX: PHIX Vx node array
- RX: RX Vx node array

Returns

- nothing

HydrologyPlanetesimals.compute_vy_node_properties! — Method

Compute properties of Vy nodes based on interpolation arrays.

compute_vy_node_properties!(RHOYSUM, RHOFYSUM, KYSUM, PHIYSUM, RYSUM, WTYSUM, RHOY, RHOFY, KY, PHIY, RY)

Details

- RHOYSUM: RHOY interpolation array
- RHOFYSUM: RHOFY interpolation array
- KYSUM: KY interpolation array
- PHIYSUM: PHIY interpolation array
- RYSUM: RY interpolation array
- WTYSUM: WTY interpolation array
- RHOY: RHOY Vy node array
- RHOFY: RHOFY Vy node array
- KY: KY Vy node array
- PHIY: PHIY Vy node array
- RY: RY Vy node array

Returns

- nothing

HydrologyPlanetesimals.define_markers! — Method

Define initial set of markers according to model parameters

define_markers!(xm, ym, tm, phim, etavpm, rhototalm, rhocptotalm, etatotalm, hrtotalm, ktotalm, tkm, inv_gggtotalm, fricttotalm, cohestotalm, tenstotalm, rhofluidcur, alphasolidcur, alphafluidcur; randomized)

Details

- xm: x coordinates of markers
- ym: y coordinates of markers
- tm: material type of markers
- phim: porosity of markers
- etavpm: matrix viscosity of markers
- rhototalm: total density of markers
- rhocptotalm: total volumetric heat capacity of markers
- etatotalm: total viscosity of markers
- hrtotalm: total radiogenic heat production of markers
- ktotalm: total thermal conductivity of markers
- tkm: temperature of markers 
- inv_gggtotalm: inverse of total shear modulus of markers
- fricttotalm: total friction coefficient of markers
- cohestotalm: total compressive strength of markers
- tenstotalm: total tensile strength of markers
- rhofluidcur: fluid density of markers
- alphasolidcur: solid thermal expansion coefficient of markers
- alphafluidcur: fluid thermal expansion coefficient of markers
- randomized: uniformly random-distribute marker x/y positions within cells
              and randomly set initial marker porosity

Returns

- nothing

HydrologyPlanetesimals.distance — Method

Calculate Euclidean distance between two point coordinates.

distance(x1, y1, x2, y2)

Details

- x1: x-coordinate of point 1 [m]
- y1: y-coordinate of point 1 [m]
- x2: x-coordinate of point 2 [m]
- y2: y-coordinate of point 2 [m]

Returns

- Euclidean distance between point 1 and point 2 [m]

HydrologyPlanetesimals.etatotal_rocks — Method

Compute total rocky marker viscosity based on temperature and material type.

etatotal_rocks(tkmm, tmm)

Details

- tkmm: marker temperature [K]
- tmm: marker type [1, 2]

Returns

- etatotal: rocky marker temperature-dependent total viscosity

HydrologyPlanetesimals.finalize_plastic_iteration_pass! — Method

Decide next pass plastic iteration time step, viscoplastic viscosity, and basic node yielding status.

finalize_plastic_iteration_pass!(ETA, ETA5, ETA00, YNY, YNY5, YNY00, YNY_inv_ETA, dt, iplast)

Details:

- ETA: viscoplastic viscosity at basic nodes
- ETA5: plastic iterations viscoplastic viscosity at basic nodes
- ETA00: previous time step viscoplastic viscosity at basic nodes
- YNY: plastic yielding status at basic nodes 
- YNY5: plastic iterations plastic yielding status at basic nodes
- YNY00: previous time step plastic yielding status at basic nodes
- YNY_inv_ETA: inverse of plastic viscosity at yielding basic nodes
- dt: current time step
- iplast: current plastic iteration counter

Returns

- dt: adjusted next time step

HydrologyPlanetesimals.fix — Method

Compute top and left grid nodes indices (i, j) and x- and y-distances to that grid node (i, j) for given (x, y) position and grid axes.

fix(x, y, x_axis, y_axis, dx, dy, jmin, jmax, imin, imax)

Details

- x: x-position [m]
- y: y-position [m]
- x_axis: x-grid reference axis array [m]
- y_axis: y-grid reference axis array [m]
- dx: x-grid axis mesh width [m]
- dy: y-grid axis mesh width [m]
- jmin: minimum assignable index on x-grid axis (basic/Vx/Vy/P)
- jmax: maximum assignable index on x-grid axis (basic/Vx/Vy/P)
- imin: minimum assignable index on y-grid axis (basic/Vx/Vy/P)
- imax: maximum assignable index on y-grid axis (basic/Vx/Vy/P)

Returns

- i: top (with reference to y) node index on y-grid axis
- j: left (with reference to x) node index on x-grid axis

HydrologyPlanetesimals.fix_distances — Method

Compute top and left grid nodes indices (i, j) and x- and y-distances to that grid node (i, j) for given (x, y) position and grid axes.

fix_distances(x, y, x_axis, y_axis, dx, dy, jmin, jmax, imin, imax)

Details

- x: x-position [m]
- y: y-position [m]
- x_axis: x-grid reference axis array [m]
- y_axis: y-grid reference axis array [m]
- dx: x-grid axis mesh width [m]
- dy: y-grid axis mesh width [m]
- jmin: minimum assignable index on x-grid axis (basic/Vx/Vy/P)
- jmax: maximum assignable index on x-grid axis (basic/Vx/Vy/P)
- imin: minimum assignable index on y-grid axis (basic/Vx/Vy/P)
- imax: maximum assignable index on y-grid axis (basic/Vx/Vy/P)

Returns

- i: top (with reference to y) node index on y-grid axis
- j: left (with reference to x) node index on x-grid axis
- dxmj: x-distance from (x, y) point to (i, j) node
- dymi: y-distance from (x, y) point to (i, j) node

HydrologyPlanetesimals.fix_weights — Method

Compute top and left grid nodes indices and bilinear interpolation weigths to nearest four grid nodes for given (x, y) position and grid axes.

fix_weights(x, y, x_axis, y_axis, dx, dy, jmin, jmax, imin, imax)

Details

- x: x-position [m]
- y: y-position [m]
- x_axis: x-grid reference axis array [m]
- y_axis: y-grid reference axis array [m]
- dx: x-grid axis mesh width [m]
- dy: y-grid axis mesh width [m]
- jmin: minimum assignable index on x-grid axis (basic/Vx/Vy/P)
- jmax: maximum assignable index on x-grid axis (basic/Vx/Vy/P)
- imin: minimum assignable index on y-grid axis (basic/Vx/Vy/P)
- imax: maximum assignable index on y-grid axis (basic/Vx/Vy/P)

Returns

- i: top (with reference to y) node index on y-grid axis
- j: left (with reference to x) node index on x-grid axis
- bilinear_weights: vector of 4 bilinear interpolation weights to
  nearest four grid nodes:
    [wtmij  : i  , j   node,
    wtmi1j : i+1, j   node,
    wtmij1 : i  , j+1 node,
    wtmi1j1: i+1, j+1 node]

HydrologyPlanetesimals.get_viscosities_stresses_density_gradients! — Method

Compute viscosities, stresses, and density gradients for hydromechanical solver.

get_viscosities_stresses_density_gradients!(ETA, ETAP, GGG, GGGP, SXY0, SXX0, RHOX, RHOY, dt, ETAcomp, ETAPcomp, SXYcomp, SXXcomp, SYYcomp, dRHOXdx, dRHOXdy, dRHOYdx, dRHOYdy)

Details

In

- ETA: viscoplastic viscosity at basic nodes
- ETAP: viscosity at P nodes
- GGG: shear modulus at basic nodes
- GGGP: shear modulus at P nodes
- SXY0: σ₀xy XY stress at basic nodes
- SXX0:σ₀xy XY stress at basic nodes
- RHOX: density at Vx nodes
- RHOY: density at Vy nodes
- dt: time step

Out

- ETAcomp: computational viscosity at basic nodes
- ETAPcomp: computational viscosity at P nodes
- SXYcomp: previous XY stresses at basic nodes
- SXXcomp: previous XX stresses at P nodes
- SYYcomp: previous YY stresses at P nodes
- dRHOXdx: density gradient at Vx nodes in x direction
- dRHOXdy: density gradient at Vx nodes in y direction
- dRHOYdx: density gradient at Vy nodes in x direction
- dRHOYdy: density gradient at Vy nodes in y direction

Returns

- nothing

HydrologyPlanetesimals.grid_average — Method

Get the arithmetic average of values from a grid 4-stencil anchored at (i, j).

grid_average(i, j, grid)

Details

- i: top left grid node column index
- j: top left grid node row index
- grid: data from which to get the average value

Returns

- grid_average: (grid[i, j]+grid[i+1, j]+grid[i, j+1]+grid[i+1, j+1]) / 4

HydrologyPlanetesimals.grid_vector — Method

Get a 4-vector of values from a grid 4-stencil anchored at (i, j) in column-major order.

grid_vector(i, j, grid)

Details

- i: top left grid node column index
- j: top left grid node row index
- grid: data from which to build Vector

Returns

- grid_vector: 4-vector of values
    [grid[i, j], grid[i+1, j], grid[i, j+1], grid[i+1, j+1]]

HydrologyPlanetesimals.interpolate_add_to_grid! — Method

Interpolate a property to neareast four nodes on a given grid location using given bilinear interpolation weights.

Details

- i: top (with reference to y) node index on vertical y-grid axis
- j: left (with reference to x) node index on horizontal x-grid axis
- weights: vector of 4 bilinear interpolation weights to
  nearest four grid nodes:
    [wtmij  : i  , j   node,
    wtmi1j : i+1, j   node,
    wtmij1 : i  , j+1 node,
    wtmi1j1: i+1, j+1 node]
- property: property to be interpolated to grid using weights
- grid: threaded grid array on which to interpolate property

Returns

- nothing

HydrologyPlanetesimals.interpolate_add_to_marker! — Method

Interpolate a property from nearest the four nodes on a given grid to a marker and add it to the markers property.

Details

- m: number of marker to interpolate to
- i: top (with reference to y) node index on vertical y-grid axis
- j: left (with reference to x) node index on horizontal x-grid axis
- weights: vector of 4 bilinear interpolation weights to
nearest four grid nodes:
    [wtmij  : i  , j   node,
    wtmi1j : i+1, j   node,
    wtmij1 : i  , j+1 node,
    wtmi1j1: i+1, j+1 node]
- marker_property: marker property array into which to interpolate and add
- property_grid: grid whose property is to be interpolated to marker
- m: marker

HydrologyPlanetesimals.interpolate_to_marker! — Method

Interpolate a property from nearest the four nodes on a given grid to a marker.

Details

- m: number of marker to interpolate to
- i: top (with reference to y) node index on vertical y-grid axis
- j: left (with reference to x) node index on horizontal x-grid axis
- weights: vector of 4 bilinear interpolation weights to
nearest four grid nodes:
    [wtmij  : i  , j   node,
    wtmi1j : i+1, j   node,
    wtmij1 : i  , j+1 node,
    wtmi1j1: i+1, j+1 node]
- marker_property: marker property array into which to interpolate
- property_grid: grid whose property is to be interpolated to marker
- m: marker

HydrologyPlanetesimals.kphi — Method

Compute porosity-dependent permeability (eqn 16.64 in Gerya (2019)).

kphi(kphim0m, phimm)

Details

- kphim0m: standard (reference) permeability (of marker type) [m^2]
- phimm: actual (marker) porosity

Returns

- kphim: empirical porosity-dependent permeability [m^2]

HydrologyPlanetesimals.ktotal — Method

Compute total thermal conductivity of two-phase material.

ktotal(ksolid, kfluid, phi)

Details

- ksolid: solid thermal conductivity [W/m/K]
- kfluid: fluid thermal conductivity [W/m/K]
- phi: fraction of solid

Returns

- ktotal: total thermal conductivity of mixed phase [W/m/K]

HydrologyPlanetesimals.marker_to_basic_nodes! — Method

Interpolate selected marker properties to basic nodes.

marker_to_basic_nodes!(m, xmm, ymm, etatotalm, etavpm, inv_gggtotalm, sxym, cohestotalm, tenstotalm, fricttotalm, ETA0SUM, ETASUM, GGGSUM, SXYSUM, COHSUM, TENSUM, FRISUM, WTSUM)

Details

- m: marker number
- xmm: marker's x-position [m]
- ymm: marker's y-position [m]
- etatotalm: total viscosity of markers
- etavpm: matrix viscosity of markers
- inv_gggtotalm: inverse of total shear modulus of markers
- sxym: marker σxy [Pa]
- cohestotalm: total compressive strength of markers
- tenstotalm: total tensile strength of markers 
- fricttotalm: total friction coefficient of markers
- ETA0SUM: viscous viscosity interpolated to basic nodes
- ETASUM: viscoplastic viscosity interpolated to basic nodes
- GGGSUM: shear modulus interpolated to basic nodes
- SXYSUM: σxy shear stress interpolated to basic nodes 
- COHSUM: compressive strength interpolated to basic nodes
- TENSUM: tensile strength interpolated to basic nodes
- FRISUM: friction interpolated to basic nodes 
- WTSUM: weight array for bilinear interpolation to basic nodes

Returns

- nothing

HydrologyPlanetesimals.marker_to_p_nodes! — Method

Interpolate selected marker properties to P nodes.

marker_to_p_nodes!(m, xmm, ymm, inv_gggtotalm, sxxm, rhototalm, rhocptotalm, alphasolidcur, alphafluidcur, hrtotalm, phim, tkm_rhocptotalm, GGGPSUM, SXXSUM, RHOSUM, RHOCPSUM, ALPHASUM, ALPHAFSUM, HRSUM, PHISUM, TKSUM, WTPSUM)

Details

- m: marker number
- xmm: marker x-position [m]
- ymm: marker y-position [m]
- inv_gggtotalm: inverse of total shear modulus of markers
- sxxm: marker σ′xx [Pa]
- rhototalm: total density of markers
- rhocptotalm: total volumetric heat capacity of markers 
- alphasolidcur: solid thermal expansion coefficient of markers 
- alphafluidcur: fluid thermal expansion coefficient of markers
- hrtotalm: total radiogenic heat production of markers
- phim:  marker porosity
- tkm_rhocptotalm: total thermal energy of markers
- GGGPSUM: shear modulus interpolated to P nodes
- SXXSUM: σ'xx interpolated to P nodes
- RHOSUM: density interpolated to P nodes
- RHOCPSUM: volumetric heat capacity interpolated to P nodes
- ALPHASUM: thermal expansion interpolated to P nodes
- ALPHAFSUM: fluid thermal expansion interpolated to P nodes
- HRSUM: radioactive heating interpolated to P nodes
- PHISUM: porosity interpolated to P nodes
- TKSUM: heat capacity interpolated to P nodes
- WTPSUM: weight for bilinear interpolation to P nodes

Returns

- nothing

HydrologyPlanetesimals.marker_to_vx_nodes! — Method

Interpolate selected marker properties to Vx nodes.

marker_to_vx_nodes!(m, xmm, ymm, rhototalm, rhofluidcur, ktotalm, phim, etafluidcur_inv_kphim, RHOXSUM, RHOFXSUM, KXSUM, PHIXSUM, RXSUM, WTXSUM)

Details

- m: marker number
- xmm: marker's x-position [m]
- ymm: marker's y-position [m]
- rhototalm: total density of markers
- rhofluidcur: fluid density of markers
- ktotalm: total thermal conductivity of markers
- phim: porosity of markers
- etafluidcur_inv_kphim: fluid viscosity over permeability of markers
- RHOXSUM: density interpolated to Vx nodes
- RHOFXSUM: fluid density interpolated to Vx nodes
- KXSUM: thermal conductivity interpolated to Vx nodes
- PHIXSUM: porosity interpolated to Vx nodes
- RXSUM: ηfluid/kϕ interpolated to Vx nodes
- WTXSUM: weight for bilinear interpolation to Vx nodes

Returns

- nothing

HydrologyPlanetesimals.marker_to_vy_nodes! — Method

Interpolate selected marker properties to Vy nodes.

marker_to_vy_nodes!(m, xmm, ymm, rhototalm, rhofluidcur, ktotalm, phim, etafluidcur_inv_kphim, RHOYSUM, RHOFYSUM, KYSUM, PHIYSUM, RYSUM, WTYSUM)

Details

- m: marker number
- xmm: marker's x-position [m]
- ymm: marker's y-position [m]
- rhototalm: total density of markers
- rhofluidcur: fluid density of markers
- ktotalm: total thermal conductivity of markers
- phim: porosity of markers
- etafluidcur_inv_kphim: fluid viscosity over permeability of markers
- RHOYSUM: density interpolated to Vy nodes
- RHOFYSUM: fluid density interpolated to Vy nodes
- KYSUM: thermal conductivity interpolated to Vy nodes
- PHIYSUM: porosity interpolated to Vy nodes
- RYSUM: ηfluid/kϕ interpolated to Vy nodes
- WTYSUM: weight for bilinear interpolation to Vy nodes

Returns

- nothing

HydrologyPlanetesimals.move_markers_rk4! — Method

Move markers using classic Runge-Kutta integration (RK4) taking into account two-phase flow and solid-fluid temperatures.

move_markers_rk4!(xm, ym, tm, tkm, phim, sxym, sxxm, vx, vy, vxf, vyf, wyx, tk2, marknum, dtm)

Details

- xm: x-coordinate of markers
- ym: y-coordinate of markers
- tm: type of markers 
- tkm: temperature of markers
- phim: porosity of markers  
- sxxm: marker σ′xx
- sxym: marker σxy
- vx: solid vx-velocity at Vx nodes
- vy: solid vy-velocity at Vy nodes
- vxf: fluid vx-velocity at Vx nodes
- vyf: fluid vy-velocity at Vy nodes
- wyx: rotation rate at basic nodes
- tk2: next temperature at P nodes
- marknum: number of markers in use
- dtm: displacement time step

Returns

- nothing

HydrologyPlanetesimals.parse_commandline — Method

Parse command line arguments and feed them to the main function.

parse_commandline()

Details:

- nothing

Returns

- parsed_args: parsed command line arguments

HydrologyPlanetesimals.perform_thermal_iterations! — Method

Perform thermal iterations to time step thermal field at P nodes.

perform_thermal_iterations!(tk0, tk1, tk2, DT, DT0, RHOCP, KX, KY, HR, HA, HS, RT, ST, dtm, timestep)

Details

- tk0: previous temperature at P nodes 
- tk1: current temperature at P nodes
- tk2: next temperature at P nodes 
- DT: calculated temperature difference at P nodes 
- DT0: previous calculated temperature difference at P nodes
- RHOCP: volumetric heat capacity at P nodes  
- KX: thermal conductivity at Vx nodes
- KY: thermal conductivity at Vy nodes 
- HR: radioactive heating at P nodes
- HA: adiabatic heating at P nodes 
- HS: shear heating at P nodes
- RT: thermal RHS coefficient vector
- ST: thermal solution vector
- dtm: displacement time step
- timestep: current time step

Returns

- nothing

HydrologyPlanetesimals.positive_max! — Method

Compare two arrays of identical sizes element-wise and fill a third array with the larger value if positive and zero otherwise.

positive_max!(A, B, C)

Details

- A: first array
- B: second array
- C: result array

Returns

- nothing

HydrologyPlanetesimals.process_hydromechanical_solution! — Method

Process hydromechanical solution vector to output physical observables.

process_hydromechanical_solution!(S, vx, vy, pr, qxD, qyD, pf)

Details

- S: hydromechanical solution vector
- vx: solid velocity at Vx nodes
- vy: solid velocity at Vy nodes
- pr: total pressure at P nodes
- qxD: qx-Darcy flux at Vx nodes
- qyD: qy-Darcy flux at Vy nodes
- pf: fluid pressure at P nodes

Returns

- nothing

HydrologyPlanetesimals.recompute_bulk_viscosity! — Method

Recompute bulk viscosity at P nodes.

Details

- ETA: viscoplastic viscosity at basic nodes
- ETAP: viscosity at P nodes
- ETAPHI: bulk viscosity at P nodes
- PHI: porosity at P nodes
- etaphikoef: coefficient: shear viscosity -> compaction viscosity

Returns

- nothing

HydrologyPlanetesimals.reduce_add_3darray! — Method

Reduce a 3D (i, j, k) along its third (k) axis by addition and write the result into (i, j, 1) without reallocating the array's memory.

reduce_add_3darray!(A)

Details

- A: 3D array [i, j, k]

Returns

- nothing

HydrologyPlanetesimals.replenish_markers! — Method

Add markers to populate currently sparsely filled grid areas.

replenish_markers!(xm, ym, tm, tkm, phim, sxxm, sxym, etavpm, rhototalm, rhocptotalm, etatotalm, hrtotalm, ktotalm, inv_gggtotalm, fricttotalm, cohestotalm, tenstotalm, rhofluidcur, alphasolidcur, alphafluidcur, tkm_rhocptotalm, etafluidcur_inv_kphim, mdis, mnum; randomized)

Details

- xm: horizontal x-position of markers
- ym: vertical y-position of markers
- tm: type of markers
- tkm: temperature of markers 
- phim: porosity of markers 
- sxxm: marker σ′xx of markers
- sxym: σxy of markers
- etavpm: viscoplastic viscosity of markers
- rhototalm: total density of markers
- rhocptotalm : total volumetric heat capacity of markers
- etatotalm: total viscosity of markers
- hrtotalm: total radiogenic heat production of markers
- ktotalm: total thermal conductivity of markers
- inv_gggtotalm: inverse of total shear modulus of markers
- fricttotalm: total friction coefficient of markers
- cohestotalm: total compressive strength of markers
- tenstotalm: total tensile strength of markers
- rhofluidcur: fluid density of markers
- alphasolidcur: solid thermal expansion coefficient of markers
- alphafluidcur: fluid thermal expansion coefficient of markers
- tkm_rhocptotalm: total thermal energy of markers
- etafluidcur_inv_kphim: fluid viscosity over permeability of markers
- mdis: minimum distance of marker launch anchor points to nearest marker 
- mnum: number of marker nearest to marker launch anchor positions

Returns

- marknum: updated number of markers in use

HydrologyPlanetesimals.reset_interpolated_properties — Method

Reset properties to be interpolated from markers to staggered grid.

reset_interpolated_properties(ETA0SUM, ETASUM, GGGSUM, SXYSUM, COHSUM, TENSUM, FRISUM, WTSUM, RHOXSUM, RHOFXSUM, KXSUM, PHIXSUM, RXSUM, WTXSUM, RHOYSUM, RHOFYSUM, KYSUM, PHIYSUM, RYSUM, WTYSUM, RHOSUM, RHOCPSUM, ALPHASUM, ALPHAFSUM, HRSUM, GGGPSUM, SXXSUM, TKSUM, PHISUM, WTPSUM)

Details

- ETA0SUM: interpolation of ETA0 at basic nodes
- ETASUM: interpolation of ETA at basic nodes
- GGGSUM: interpolation of GGG at basic nodes
- SXYSUM: interpolation of SXY at basic nodes
- COHSUM: interpolation of COH at basic nodes
- TENSUM: interpolation of TEN at basic nodes
- FRISUM: interpolation of FRI at basic nodes
- WTSUM: interpolation weights at basic nodes
- RHOXSUM: interpolation of RHOX at Vx nodes
- RHOFXSUM: interpolation of RHOFX at Vx nodes
- KXSUM: interpolation of KX at Vx nodes
- PHIXSUM: interpolation of PHIX at Vx nodes
- RXSUM: interpolation of RX at Vx nodes
- WTXSUM: interpolation weights at Vx nodes
- RHOYSUM: interpolation of RHOY at Vy nodes
- RHOFYSUM: interpolation of RHOFY at Vy nodes
- KYSUM: interpolation of KY at Vy nodes
- PHIYSUM: interpolation of PHIX at Vy nodes
- RYSUM: interpolation of RY at Vy nodes
- WTYSUM: interpolation weights at Vy nodes
- RHOSUM: interpolation of RHO at P nodes
- RHOCPSUM: interpolation of RHOCP at P nodes
- ALPHASUM: interpolation of ALPHA at P nodes
- ALPHAFSUM: interpolation of ALPHAF at P nodes
- HRSUM: interpolation of HR at P nodes
- GGGPSUM: interpolation of GGGP at P nodes
- SXXSUM: interpolation of SXX at P nodes
- TKSUM: interpolation of TK at P nodes
- PHISUM: interpolation of PHI at P nodes
- WTPSUM: interpolation weights at P nodes

Returns

- nothing

HydrologyPlanetesimals.run_simulation — Method

Runs the simulation with the given parameters.

run_simulation()

Details

- nothing

Returns

- nothing

HydrologyPlanetesimals.save_state — Method

Save simulation state to JLD2 output file named after current timestep.

save_state(output_path, timestep, dt, dtm, timesum, marknum, ETA, ETA0, GGG, EXY, SXY, SXY0, wyx, COH, TEN, FRI, YNY, RHOX, RHOFX, KX, PHIX, vx, vxf, RX, qxD, gx, RHOY, RHOFY, KY, PHIY, vy, vyf, RY, qyD, gy, RHO, RHOCP, ALPHA, ALPHAF, HR, HA, HS, ETAP, GGGP, EXX, SXX, SXX0, tk1, tk2, vxp, vyp, vxpf, vypf, pr, pf, ps, pr0, pf0, ps0, ETAPHI, BETTAPHI, PHI, APHI, FI, ETA5, ETA00, YNY5, YNY00, YNY_inv_ETA, DSXY, EII, SII, DSXX, xm, ym, tm, tkm, sxxm, sxym, etavpm, phim, rhototalm, rhocptotalm, etatotalm, hrtotalm, ktotalm, tkm_rhocptotalm, etafluidcur_inv_kphim, inv_gggtotalm, fricttotalm, cohestotalm, tenstotalm, rhofluidcur, alphasolidcur, alphafluidcur)

Details

- output_path: absolute path to output directory
- timestep: current time step number
- dt: time step
- dtm: displacement time step 
- timesum: total simulation time
- marknum: number of markers
- ETA... : simulation state variables

Returns

- nothing

HydrologyPlanetesimals.setup_dynamic_simulation_parameters — Method

Set up and initialize dynamic simulation parameters.

setup_dynamic_simulation_parameters()

Details

- nothing

Returns

- timestep: simulation starting time step count
- dt: simulation initial computational time step [s]
- timesum: simulation starting time [s]
- marknum: initial number of markers
- hrsolidm: initial radiogenic heat production solid phase
- hrfluidm: initial radiogenic heat production fluid phase
- YERRNOD: vector of summed yielding errors of nodes over plastic iterations

HydrologyPlanetesimals.setup_gravitational_lse — Method

Set up gravitational linear system of equations structures.

setup_gravitational_lse()

Details

- nothing

Returns

- RP: gravitational linear system of equations: RHS vector
- SP: gravitational linear system of equations: solution vector

HydrologyPlanetesimals.setup_hydromechanical_lse — Method

Set up hydromechanical linear system of equations structures.

setup_hydromechanical_lse()

Details

- nothing

Returns

- R: hydromechanical linear system of equations: RHS vector
- S: hydromechanical linear system of equations: solution vector

HydrologyPlanetesimals.setup_interpolated_properties — Method

Set up properties to be interpolated from markers to staggered grid.

setup_interpolated_properties()

Details

- nothing

Returns

- ETA0SUM: interpolation of ETA0 at basic nodes
- ETASUM: interpolation of ETA at basic nodes
- GGGSUM: interpolation of GGG at basic nodes
- SXYSUM: interpolation of SXY at basic nodes
- COHSUM: interpolation of COH at basic nodes
- TENSUM: interpolation of TEN at basic nodes
- FRISUM: interpolation of FRI at basic nodes
- WTSUM: interpolation weights at basic nodes
- RHOXSUM: interpolation of RHOX at Vx nodes
- RHOFXSUM: interpolation of RHOFX at Vx nodes
- KXSUM: interpolation of KX at Vx nodes
- PHIXSUM: interpolation of PHIX at Vx nodes
- RXSUM: interpolation of RX at Vx nodes
- WTXSUM: interpolation weights at Vx nodes
- RHOYSUM: interpolation of RHOY at Vy nodes
- RHOFYSUM: interpolation of RHOFY at Vy nodes
- KYSUM: interpolation of KY at Vy nodes
- PHIYSUM: interpolation of PHIX at Vy nodes
- RYSUM: interpolation of RY at Vy nodes
- WTYSUM: interpolation weights at Vy nodes
- RHOSUM: interpolation of RHO at P nodes
- RHOCPSUM: interpolation of RHOCP at P nodes
- ALPHASUM: interpolation of ALPHA at P nodes
- ALPHAFSUM: interpolation of ALPHAF at P nodes
- HRSUM: interpolation of HR at P nodes
- GGGPSUM: interpolation of GGGP at P nodes
- SXXSUM: interpolation of SXX at P nodes
- TKSUM: interpolation of TK at P nodes
- PHISUM: interpolation of PHI at P nodes
- WTPSUM: interpolation weights at P nodes

HydrologyPlanetesimals.setup_marker_geometry_helpers — Method

Set up additional marker geometry helpers to facilitate marker handling.

setup_marker_geometry_helpers()

Details

- nothing

Returns

- mdis: minimum distance of marker launch anchor points to nearest marker
- mnum: number of marker nearest to marker launch anchor positions

HydrologyPlanetesimals.setup_marker_properties — Method

Set up geodesic and physical properties of the set of markers.

setup_marker_properties(marknum; randomized)

Details

- randomized: fill in random values for grid properties instead of zeros

Returns

- xm : horizontal marker coordinate [m]
- ym : vertical marker coordinate [m]
- tm : marker material type
- tkm : marker temperature [K]
- sxxm : marker σ′xx [Pa]
- sxym : marker σxy [Pa]
- etavpm : marker viscoplastic viscosity [Pa]
- phim : marker porosity

HydrologyPlanetesimals.setup_marker_properties_helpers — Method

Set up additional helper marker properties to facility comptuations.

setup_marker_properties_helpers(marknum; randomized)

Details

- randomized: fill in random values for grid properties instead of zeros

Returns

- rhototalm: total density of markers
- rhocptotalm : total volumetric heat capacity of markers
- etatotalm: total viscosity of markers
- hrtotalm: total radiogenic heat production of markers
- ktotalm: total thermal conductivity of markers
- tkm_rhocptotalm: total thermal energy of markers
- etafluidcur_inv_kphim: fluid viscosity over permeability of markers
- inv_gggtotalm: inverse of total shear modulus of markers
- fricttotalm: total friction coefficient of markers
- cohestotalm: total compressive strength of markers
- tenstotalm: total tensile strength of markers
- rhofluidcur: fluid density of markers
- alphasolidcur: solid thermal expansion coefficient of markers
- alphafluidcur: fluid thermal expansion coefficient of markers

HydrologyPlanetesimals.setup_staggered_grid_properties — Method

Set up staggered grid properties for basic, Vx, Vy, and P nodes.

setup_staggered_grid_properties(; randomized)

Details

- randomized: fill in random values for grid properties instead of zeros

Returns

- ETA : viscoplastic viscosity at basic nodes [Pa*s]
- ETA0 : viscous viscosity at basic nodes [Pa*s]
- GGG : shear modulus at basic nodes [Pa]
- EXY : ϵxy at basic nodes [1/s]
- SXY : σxy at basic nodes [1/s]
- SXY0 : σ₀xy at basic nodes [1/s]
- wyx : rotation rate at basic nodes [1/s]
- COH : compressive strength at basic nodes [Pa]
- TEN : tensile strength at basic nodes [Pa]
- FRI : friction at basic nodes
- YNY : plastic yielding node property at basic nodes
- RHOX : density at Vx nodes [kg/m^3]
- RHOFX : fluid density at Vx nodes [kg/m^3]
- KX : thermal conductivity at Vx nodes [W/m/K]
- PHIX : porosity at Vx nodes
- vx : solid vx-velocity at Vx nodes [m/s]
- vxf : fluid vx-velocity at Vx nodes [m/s]
- RX : etafluid/kphi ratio at Vx nodes [m^2]
- qxD : qx-darcy flux at Vx nodes [m/s]
- gx : gx-gravity at Vx nodes [m/s^2]
- RHOY : density at Vy nodes [kg/m^3]
- RHOFY : fluid density at Vy nodes [kg/m^3]
- KY : thermal conductivity at Vy nodes [W/m/K]
- PHIY : porosity at Vy nodes
- vy : solid vy-velocity at Vy nodes [m/s]
- vyf : fluid vy-velocity at Vy nodes [m/s]
- RY : etafluid/kphi ratio at Vy nodes [m^2]
- qyD : qy-Darcy flux at Vy nodes [m/s]
- gy : gy-gravity at Vy nodes [m/s^2]
- RHO : density at P nodes [kg/m^3]
- RHOCP : volumetric heat capacity at P nodes [J/m^3/K]
- ALPHA : thermal expansion at P nodes [J/m^3/K]
- ALPHAF : fluid thermal expansion at P nodes [J/m^3/K]
- HR : radioactive heating at P nodes [W/m^3]
- HA : adiabatic heating at P nodes [W/m^3]
- HS : shear heating at P nodes [W/m^3]
- ETAP : viscosity at P nodes [Pa*s]
- GGGP : shear modulus at P nodes [Pa]
- EXX : ϵxx at P nodes [1/s]
- SXX : σ′xx at P nodes [1/s]
- SXX0 : σ₀′xx at P nodes [1/s]
- tk1 : current temperature at P nodes [K]
- tk2 : next temperature at P nodes [K]
- DT: temperature difference at P nodes [K]
- DT0: previous temperature difference at P nodes [K]
- vxp : solid vx in pressure nodes at P nodes [m/s]
- vyp : solid vy in pressure nodes at P nodes [m/s]
- vxpf : fluid vx in pressure nodes at P nodes [m/s]
- vypf : fluid vy in pressure nodes at P nodes [m/s]
- pr : total pressure at P nodes [Pa]
- pf : fluid pressure at P nodes [Pa]
- ps : solid pressure at P nodes [Pa]
- pr0 : previous total pressure at P nodes [Pa]
- pf0 : previous fluid pressure at P nodes [Pa]
- ps0 : previous solid pressure at P nodes [Pa]
- ETAPHI : bulk viscosity at P nodes [Pa*s]
- BETTAPHI : bulk compresibility at P nodes [Pa*s]
- PHI : porosity at P nodes
- APHI : Dlnat P nodes [(1-ϕ)/ϕ]/Dt
- FI : gravity potential at P nodes [J/kg]

HydrologyPlanetesimals.setup_staggered_grid_properties_helpers — Method

Set up additional helper staggered grid properties to facilitate computations.

setup_staggered_grid_properties_helpers(; randomized)

Details

- randomized: fill in random values for grid properties instead of zeros

Returns

- ETA5: plastic iterations viscoplastic viscosity at basic nodes [Pa⋅s]
- ETA00: previous viscous viscosity at basic nodes [Pa⋅s]
- YNY5: plastic iterations plastic yielding node property at basic nodes
- YNY00: previous plastic yielding node property at basic nodes
- DSXY: stress change Δσxy at basic nodes [Pa]
- DSY: (SIIB-syield) at basic nodes
- EII :second strain rate invariant at P nodes [1/s]
- SII :second stress invariant at P nodes [Pa]
- DSXX :stress change Δσ′xx at P nodes [Pa]
- tk0: previous temperature at P nodes [K]

HydrologyPlanetesimals.setup_thermal_lse — Method

Set up thermal linear system of equations structures.

setup_thermal_lse()

Details

- nothing

Returns

- RT: thermal linear system of equations: RHS vector
- ST: thermal linear system of equations: solution vector

HydrologyPlanetesimals.simulation_loop — Method

Main simulation loop: run calculations with timestepping.

simulation_loop(output_path)

Details

- output_path: Absolute path where to save simulation output files

Returns

- nothing

HydrologyPlanetesimals.symmetrize_p_node_observables! — Method

Apply symmetry to P node observables.

symmetrize_p_node_observables!(SXX, APHI, PHI, pr, pf, ps)

Details

- SXX: σ′xx at P nodes
- APHI: Aϕ = Dln[(1-ϕ)/ϕ]/Dt at P nodes
- PHI: porosity at P nodes
- pr: total pressure at P nodes
- pf: fluid pressure at P nodes
- ps: solid pressure at P nodes

Returns

- nothing

HydrologyPlanetesimals.total — Method

Compute convex combination of fluid and solid properties to get total property.

total(solid, fluid, ϕ)

Details

- fluid: fluid properties
- solid: solid properties
- ϕ: porosity (fraction of fluid)

Returns

- total: computed total property

HydrologyPlanetesimals.update_marker_population_geometry! — Method

Update marker population geometry status given a marker number and nearest top/left marker grid point.

update_marker_population_geometry!(m, i, j, xm, ym, mdis, mnum)

Details

- m: marker number
- i: vertical index of top/left marker grid point
- j: horizontal index of top/left marker grid point
- xm: horizontal x-position of markers
- ym: vertical y-position of markers
- mdis: minimum distance of marker launch anchor points to nearest marker
- mnum: number of marker nearest to marker launch anchor positions

Returns

- nothing

HydrologyPlanetesimals.update_marker_porosity! — Method

Update marker porosity based on Dln[(1-ϕ)/ϕ]/Dt at P grid.

update_marker_porosity!(xm, ym, tm, phim, APHI, dtm, marknum)

Details

- xm: x-coordinates of markers
- ym: y-coordinates of markers
- tm: marker type
- phim: marker porosity
- APHI: Dln[(1-ϕ)/ϕ]/Dt at P nodes
- dtm: displacement time step
- marknum: total number of markers in use

Returns

- nothing

HydrologyPlanetesimals.update_marker_stress! — Method

Update marker stress based on xy (basic grid) and xx (P grid) stress changes.

update_marker_stress!(xm, ym, sxxm, sxym, DSXX, DSXY, marknum)

Details

- xm: x-coordinates of markers
- ym: y-coordinates of markers
- sxxm: marker σ′xx [Pa]
- sxym: marker σxy [Pa]
- DSXX: stress change Δσ′xx at P nodes
- DSXY: stress change Δσxy at basic nodes
- marknum: total number of markers in use

Returns

- nothing

HydrologyPlanetesimals.update_marker_temperature! — Method

Update marker temperature based on P grid temperature changes.

update_marker_temperature!(xm, ym, tkm, DT, tk2, timestep, marknum)

Details

- xm: x-coordinates of markers
- ym: y-coordinates of markers
- tkm: marker temperature
- DT: temperature change at P nodes
- tk2: next temperature at P nodes
- timestep: current time step
- marknum: total number of markers in use

Returns

- nothing

HydrologyPlanetesimals.update_marker_viscosity! — Method

Update marker viscoplastic viscosity based on basic node yielding status and marker temperature- and material-based viscosity.

update_marker_viscosity!(m, xm, ym, tm, tkm, etatotalm, etavpm, YNY, YNY_inv_ETA)

Details

- m: marker number
- xm: x-coordinates of markers
- ym: y-coordinates of markers
- tm: type of markers
- tkm: temperature of markers
- etatotalm: total viscosity of markers
- etavpm: matrix viscosity of markers
- YNY: plastic yielding node property at basic nodes
- YNY_inv_ETA: inverse viscoplastic viscosity at yielding basic nodes

Returns

-nothing

HydrologyPlanetesimals.ηᶠcur_inv_kᵠ — Method

Compute inverse of porosity-dependent permeability (eqn 16.64 in Gerya (2019)) times current fluid viscosity.

ηᶠcur_inv_kᵠ(kϕᵣ, ϕ, ηᶠcur)

Details

- kϕᵣ: reference permeability [m^2]
- ϕ: current porosity
- ηᶠcur: current fluid viscosity [Pa s]

Returns

- etafluidcur_inv_kphi: inverse empirical porosity-dependent permeability 
                        times current fluid viscosity